requirements.txt
Flask==1.0.2
Flask-PyMongo==2.2.0
pymongo==3.7.2
python-dotenv==0.10.1
applicationinsights==0.11.7